Modularize input handling
