+++ src/ui/input/mod.rs
@@ -1,0 +1,5 @@
+pub mod keyboard;
+pub mod mouse;
+
+pub use keyboard::handle_key;
+pub use mouse::handle_mouse;
